HostDrivers/Support: build fix.
svn:sync-xref-src-repo-rev: r96764

+++ src/VBox/HostDrivers/Support/SUPDrv.c
@@ -1,4 +1,4 @@
-/* $Id: SUPDrv.c 53215 2014-11-04 18:44:59Z ramshankar.venkataraman@oracle.com $ */
+/* $Id: SUPDrv.c 53216 2014-11-04 18:47:26Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  * VBoxDrv - The VirtualBox Support Driver - Common code.
  */
@@ -7447,7 +7447,7 @@ static void supdrvGipUpdate(PSUPDRVDEVEXT pDevExt, uint64_t u64NanoTS, uint64_t 
                  *        calculation on non-invariant hosts if it changes the history decision
                  *        taken in supdrvGipDoUpdateCpu(). */
                 ASMAtomicWriteU32(&pGip->u32UpdateHz, u32UpdateHz);
-                ASMAtomicWriteU32(&pGip->u32UpdateIntervalNS, u64Delta / GIP_UPDATEHZ_RECALC_FREQ);
+                ASMAtomicWriteU32(&pGip->u32UpdateIntervalNS, (uint32_t)u64Delta / GIP_UPDATEHZ_RECALC_FREQ);
             }
 #endif
         }
@@ -7629,7 +7629,7 @@ static int supdrvIOCtl_TscRead(PSUPDRVDEVEXT pDevExt, PSUPTSCREAD pReq)
 {
     uint64_t uTsc;
     uint16_t idApic;
-    int16_t cTries;
+    int16_t  cTries;
     PSUPGLOBALINFOPAGE pGip;
     int rc;
